pyproject.toml
[tool.poetry]
name = "interview-example"
version = "0.1.0"
description = ""
authors = ["Андрей Солдатов <AFeSoldatov@sberbank.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
redis = "^5.2.1"
django-redis = "^5.4.0"
djangorestframework = "^3.16.0"
psycopg = "^3.2.6"
pydantic-settings = "^2.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
